services/fundamentals/requirements.txt
fastapi>=0.115.0
uvicorn>=0.34.0
httpx>=0.27.0
redis>=5.0.0
pyyaml>=6.0
pandas>=2.0.0
numpy>=1.24.0
